FOV was an int, lol

+++ includes/camera.h
@@ -6,7 +6,7 @@
 /*   By: gcassi-d <gcassi-d@42urduliz.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/14 21:45:11 by gcassi-d          #+#    #+#             */
-/*   Updated: 2026/02/20 19:04:30 by gcassi-d         ###   ########.fr       */
+/*   Updated: 2026/02/21 12:23:48 by gcassi-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ typedef struct s_camera
 	t_vec	dir;
 	t_vec	right;
 	t_vec	up;
-	double	fov;
+	int		fov;
 	char	isdef;
 }	t_camera;
